Flatten app interface
