+++ src/api/persistence/mongoose/index.test.ts
@@ -2,7 +2,7 @@ import dbConnect, { close, schema } from '.';
 import loadConfig from '../../../config';
 
 jest.mock('./connect', () => ({
-  _esModule: true,
+  __esModule: true,
   default: jest.fn().mockImplementation(
     (mongoose, config) =>
       new Promise((resolve, reject) => {
@@ -14,15 +14,17 @@ jest.mock('./connect', () => ({
 }));
 
 jest.mock('./queries', () => ({
-  _esModule: true,
+  __esModule: true,
   default: jest.fn().mockImplementation(client => (client ? true : false)),
 }));
 
 describe('mongoose', () => {
   const mongo = {
     disconnect: jest.fn().mockResolvedValue(true),
-  };
-  const config = loadConfig(null, null, null);
+  } as any;
+
+  const config = loadConfig('', {});
+
   describe('.connect', () => {
     it('connects to mongodb', async () => {
       const mongoose = await dbConnect(config);
