+++ src/env.ts
@@ -1,3 +1,3 @@
 import dotenv from 'dotenv';
 
-dotenv.config({ silent: true });
+dotenv.config({ debug: true });
